+++ Core/Gen5/Searchers/ProfileSearcher5.hpp
@@ -36,7 +36,7 @@ class ProfileSearcher5
 {
 public:
     ProfileSearcher5() = default;
-    explicit ProfileSearcher5(const Date &date, const Time &time, int minSeconds, int maxSeconds, u8 minVCount, u8 maxVCount, u16 minTimer0,
+    explicit ProfileSearcher5(const Date &date, const Time &time, u8 minSeconds, u8 maxSeconds, u8 minVCount, u8 maxVCount, u16 minTimer0,
                               u16 maxTimer0, u8 minGxStat, u8 maxGxStat, bool softReset, Game version, Language language, DSType dsType,
                               u64 mac, Buttons keypress);
     virtual ~ProfileSearcher5() = default;
@@ -48,8 +48,8 @@ public:
 private:
     Date date;
     Time time;
-    int minSeconds;
-    int maxSeconds;
+    u8 minSeconds;
+    u8 maxSeconds;
     u8 minVCount;
     u8 maxVCount;
     u16 minTimer0;
